+++ CHANGES.rst
@@ -26,6 +26,8 @@ Unreleased
 * **[BUGFIX]** Fixed methods which require more than one OAuth scope.
 * **[BUGFIX]** Fixed :meth:`praw.objects.WikiPage.remove_editor` raising
   AssertionError when used through OAuth.
+* **[CHANGE]** :meth:`praw.objects.Refreshable.refresh` will now always return
+  a fresh object. Previously it would use cache content when available.
 * **[FEATURE]** Added methods :meth:`leave_moderator` and
   :meth:`leave_contributor` to :class:`praw.__init__.AuthenticatedReddit`
   and :class:`praw.objects.Subreddit`.